refact: restructure project packages

+++ src/calculator/pot_splitter.py
@@ -1,12 +1,12 @@
-from model.player import Player
-from model.bet import Bet
+from src.model.player import Player
+from src.model.bet import Bet
 
 class PotSplitter:
     def calculate_pot_distribution(players: list[Player], champion: str) -> list[Player]:
-        absolute_pot: int = sum(player.bet.value for player in players)
+        absolute_pot: float = sum(player.bet.value for player in players)
 
         winners: list[Player] = list(filter(lambda player: player.bet.target == champion, players))
-        winners_pot: int = sum(player.bet.value for player in winners)
+        winners_pot: float = sum(player.bet.value for player in winners)
 
         distribution: list[Player] = [
             Player(
